requirements.txt
jinja2
mysql-connector-python
pytest
sqlparse